+++ custom_modules/hr_contract_psi/views/hr_contract_views.xml
@@ -13,63 +13,6 @@
 		</field>
 	</record>
 
-	<record id="hr_hr_employee_view_form_psi" model="ir.ui.view">
-		<field name="name">hr.hr.employee.view.form.psi</field>
-		<field name="model">hr.employee</field>
-		<field name="inherit_id" ref="hr.view_employee_form" />
-		<field name="arch" type="xml">
-			<div name="button_box" position="inside">
-				<button name="action_get_attachment_tree_view" class="oe_stat_button"
-					icon="fa-book" type="object">
-					<field name="attachment_number" widget="statinfo" string="Documents"
-						modifiers="{'readonly': true}" />
-				</button>
-			</div>
-			
-			<xpath expr="//field[@name='type_id']" position='replace'>
-           		<field name="type_id" invisible="1"/>
-        	</xpath> 
-			
-			<xpath expr="//field[@name='type_id']" position='after'>
-           		<field name="psi_contract_type"/>
-           		<field name="place_of_work"/>
-        	</xpath>
-			
-			<page name="public" position="after">
-				<page string="Pièces à fournir">
-					<group>
-						<field name="personal_information" />
-						<field name="id_photos" />
-						<field name="certificate_residence" />
-						<field name="marriage_certificate" />
-						<field name="cin" />
-						<field name="work_certificate" />
-						<field name="qualifications" />
-						<field name="criminal_records" />
-						<field name="card_cnaps" />
-						<field name="birth_certificate_children" />
-						<field name="ethics_course_certificate" />
-					</group>
-				</page>
-			</page>
-			<page name="hr_settings" position="after">
-				<page string="Sanctions">
-					<group>
-						<field name="sanctions_data" nolabel="1">
-							<tree string="" editable="bottom">
-								<field name="sanction_date" />
-								<field name="sanction_type_id" />
-								<field name="sanction_objet" />
-								<field name="sanction_date_effacement" />
-								<field name="sanction_commentaire" />
-							</tree>
-						</field>
-					</group>
-				</page>
-			</page>
-		</field>
-	</record>
-
 	<!-- Resume and Letters -->
 	<record id="hr_contract.ir_attachment_view_search_inherit_hr_employee"
 		model="ir.ui.view">
@@ -93,6 +36,7 @@
 		</field>
 	</record>
 
+	<!-- TODO: What is this ??? -->
 	<record model="ir.actions.act_window" id="hr_contract_resumes">
 		<field name="name">Resumes and Letters</field>
 		<field name="res_model">ir.attachment</field>
@@ -155,7 +99,6 @@
 		</field>
 	</record>
 
-
 	<record id="open_contract_wage_grid" model="ir.actions.act_window">
 		<field name="name">Grille salariale</field>
 		<field name="res_model">hr.wage.grid</field>
@@ -164,9 +107,7 @@
 		<field name="view_id" ref="hr_contract_wage_grid_form" />
 	</record>
 
-	
-
-  <record id="hr_contract_type_view_form" model="ir.ui.view">
+  	<record id="hr_contract_type_view_form" model="ir.ui.view">
          <field name="name">hr.contract.type.sanction.form</field>
          <field name="model">hr.contract.type.sanction</field>
          <field name="arch" type="xml">
