+++ custom_modules/hr_employee_psi/data/hr_employee_cron.xml
@@ -1,0 +1,45 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+	<data noupdate="1">
+		<!-- Cron send email collaborateur and RH - Suivi date d’anniversaire - rappel -->
+        <record id="ir_cron_suivi_date_d_anniversaire" forcecreate="True" model="ir.cron">
+            <field name="name">Suivi date d'anniversaire rappel a Collaborateur et RH</field>
+            <field name="user_id" ref="base.user_root" />
+            <field name="interval_number">1</field>
+            <field name="interval_type">minutes</field>
+            <field name="numbercall">-1</field>
+            <field name="doall">1</field>
+            <field eval="'hr.employee'" name="model" />
+            <field name="function" eval="'_send_email_birthday_date_tracking'" /> 
+            <field name="args" eval=""/>
+            <field name="active" eval="True" />
+        </record>
+        
+        <!-- Cron send second email collaborator -->
+        <record id="ir_cron_send_email_collab_2" model="ir.cron">
+            <field name="name">Collaborateur - Second rappel au collaborateur</field>
+            <field name="active" eval="False" />
+            <field name="user_id" ref="base.user_root" />
+            <field name="interval_number">2</field>
+            <field name="interval_type">minutes</field>	<!-- 15 days -->
+            <field name="numbercall">1</field>
+            <field name="doall" eval="False" />
+            <field eval="'hr.employee'" name="model" />
+            <field name="function" eval="'_send_second_email_collaborator'" /> 
+        </record>
+        
+        <!-- Cron send first email collaborator -->
+        <record id="ir_cron_send_email_collab_1" forcecreate="True" model="ir.cron">
+            <field name="name">Collaborateur - Premier rappel Email au collaborateur</field>
+            <field name="user_id" ref="base.user_root" />
+            <field name="interval_number">1</field>
+            <field name="interval_type">minutes</field> <!-- 15 days -->
+            <field name="numbercall">1</field>
+            <field name="doall" eval="False" />
+            <field eval="'hr.employee'" name="model" />
+            <field name="function" eval="'_send_first_email_collaborator'" /> 
+            <field name="args" eval="'(True,)'"/>
+            <field name="active" eval="False" />
+        </record>     
+    </data>
+</odoo>
