+++ custom_modules/hr_employee_psi/data/hr_employee_data.xml
@@ -1,18 +1,22 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
 	<data noupdate="1">
-		<!-- Cron send email collaborateur and RH - Suivi date d’anniversaire - rappel -->
-        <record id="ir_cron_suivi_date_d_anniversaire" forcecreate="True" model="ir.cron">
-            <field name="name">Suivi date d'anniversaire rappel a Collaborateur et RH</field>
-            <field name="user_id" ref="base.user_root" />
-            <field name="interval_number">1</field>
-            <field name="interval_type">minutes</field>
-            <field name="numbercall">-1</field>
-            <field name="doall">1</field>
-            <field eval="'hr.employee'" name="model" />
-            <field name="function" eval="'_send_email_birthday_date_tracking'" /> 
-            <field name="args" eval=""/>
-            <field name="active" eval="True" />
-        </record>     
+		<!-- Liste des sanctions -->
+		<record model="hr.contract.type.sanction" id="rappel_ordre">
+	        <field name="name">Rappel à l’ordre</field>
+	    </record>
+	    <record model="hr.contract.type.sanction" id="sanction_legere">
+	        <field name="name">Sanctions légères</field>
+	    </record>
+	    <record model="hr.contract.type.sanction" id="sanction_severe">
+	        <field name="name">Sanctions sévères</field>
+	    </record>
+	    <record model="hr.contract.type.sanction" id="sanction_grave">
+	        <field name="name">Sanctions graves</field>
+	    </record>
+	    <record model="hr.contract.type.sanction" id="faute_lourde">
+	        <field name="name">Fautes lourdes</field>
+	    </record>
+		<!-- Fin liste sanctions -->     
     </data>
 </odoo>
